// pw_web/log-viewer/src/legacy/models/log_data.ts
// Copyright 2023 The Pigweed Authors
//
// Licensed under the Apache License, Version 2.0 (the "License"); you may not
// use this file except in compliance with the License. You may obtain a copy of
// the License at
//
//     https://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
// License for the specific language governing permissions and limitations under
// the License.

/* eslint-disable @typescript-eslint/naming-convention */
export interface FfxLogData {
    data: PossibleDatas;
    timestamp: number;
    version: number;
}

interface PossibleDatas {
    TargetLog?: LogData;
    FfxEvent?: FfxEventData;
    MalformedTargetLog?: string;
    SymbolizedTargetLog?: [LogData, string];
    ViewerEvent?: string;
}

export type FfxEventData = "TargetDisconnected" | "LoggingStarted";

export interface LogData {
    metadata: LogMetadata;
    moniker: string;
    payload: LogPayload;
}

export interface LogMetadata {
    errors: Array<unknown>;
    component_url: string;
    timestamp: number;
    severity: string;
    tags: Array<string> | undefined | null;
    pid: number | undefined | null;
    tid: number | undefined | null;
    file: string | undefined | null;
    line: number | undefined | null;
}

export interface LogPayload {
    root: PayloadRoot;
}

interface PayloadRoot {
    message: LogMessage | undefined | null;
    keys: Record<string, unknown> | undefined | null;
    printf: LogPrintf | undefined | null;
}

interface LogMessage {
    value: string;
}

interface LogPrintf {
    format: string;
    args: Array<unknown>;
}
export interface LogRowData {
    fields: Array<LogRowValues>;
    timestamp?: number;
}

export interface LogRowValues {
    key: string;
    text: string;
    dataFields?: Array<LogRowDataField>;
}

export interface LogRowDataField {
    key: string;
    value: unknown;
}